+++ CHANGELOG.md
@@ -14,11 +14,16 @@
 
 ## [Unreleased]
 ### Added    
+### Changed    
+### Removed    
+
+## [0.9.0] - 2024-06-24
+### Added    
 * `StereoSquare2D` and `StereoSquare3D` classes for making a 2x2 grid showing all stereogram types at once - side-by-side, anaglyph, and wiggle
 * Can specify a target axis for wiggle stereograms
+* Numpy 2.0 compatibility
 ### Changed    
 * Fix `Axes3D.plot()` and `.plot3D()` not working for 3D plotting
-### Removed    
 
 ## [0.8.1] - 2024-02-19
 Note: v0.8.0 never published to PyPI, skipped.